V3 migration: security review (#210)
* settle savings interest under old mode before switching

* widen challenge number param from uint32 to uint256

* expand fuzz testing with savings, equity, bridge & governance coverage

* add fuzzing to CI pipeline

* update README

* cleanup

* fix CI tests

* internal audit

+++ contracts/MintingHubV3/MintingHub.sol
@@ -20,6 +20,10 @@ import {PositionRoller} from "./PositionRoller.sol";
  * @notice The central hub for creating, cloning, and challenging collateralized DecentralizedEURO positions.
  * @dev Only one instance of this contract is required, whereas every new position comes with a new position
  * contract. Pending challenges are stored as structs in an array.
+ *
+ * Unsupported collateral token types (enforced via governance):
+ * - Fee-on-transfer tokens: break collateral accounting (actual balance < recorded amount)
+ * - Rebasing tokens: break challenge accounting (challengedAmount becomes stale after rebase)
  */
 contract MintingHub is IMintingHub, ERC165, Leadrate {
     /**
@@ -267,18 +271,18 @@ contract MintingHub is IMintingHub, ERC165, Leadrate {
      *                                  (automatically reduced to the available amount)
      * @param postponeCollateralReturn  To postpone the return of the collateral to the challenger. Usually false.
      */
-    function bid(uint32 _challengeNumber, uint256 size, bool postponeCollateralReturn, bool returnCollateralAsNative) external {
+    function bid(uint256 _challengeNumber, uint256 size, bool postponeCollateralReturn, bool returnCollateralAsNative) external {
         _bid(_challengeNumber, size, postponeCollateralReturn, returnCollateralAsNative);
     }
 
     /**
      * @notice Post a bid in deur given an open challenge (backward compatible).
      */
-    function bid(uint32 _challengeNumber, uint256 size, bool postponeCollateralReturn) external {
+    function bid(uint256 _challengeNumber, uint256 size, bool postponeCollateralReturn) external {
         _bid(_challengeNumber, size, postponeCollateralReturn, false);
     }
 
-    function _bid(uint32 _challengeNumber, uint256 size, bool postponeCollateralReturn, bool returnCollateralAsNative) internal {
+    function _bid(uint256 _challengeNumber, uint256 size, bool postponeCollateralReturn, bool returnCollateralAsNative) internal {
         Challenge memory _challenge = challenges[_challengeNumber];
         (uint256 liqPrice, uint40 phase) = _challenge.position.challengeData();
         size = _challenge.size < size ? _challenge.size : size; // cannot bid for more than the size of the challenge
@@ -334,7 +338,7 @@ contract MintingHub is IMintingHub, ERC165, Leadrate {
         return (collateral, offer);
     }
 
-    function _avertChallenge(Challenge memory _challenge, uint32 number, uint256 liqPrice, uint256 size, bool asNative) internal {
+    function _avertChallenge(Challenge memory _challenge, uint256 number, uint256 liqPrice, uint256 size, bool asNative) internal {
         require(block.timestamp != _challenge.start); // do not allow to avert the challenge in the same transaction, see CS-ZCHF-037
         if (msg.sender == _challenge.challenger) {
             // allow challenger to cancel challenge without paying themselves
@@ -363,7 +367,7 @@ contract MintingHub is IMintingHub, ERC165, Leadrate {
      */
     function _returnChallengerCollateral(
         Challenge memory _challenge,
-        uint32 number,
+        uint256 number,
         uint256 amount,
         bool postpone,
         bool asNative
@@ -408,7 +412,7 @@ contract MintingHub is IMintingHub, ERC165, Leadrate {
      * @dev The price comes with (36 - collateral.decimals()) digits, so multiplying it with the raw collateral amount
      * always yields a price with 36 digits, or 18 digits after dividing by 10**18 again.
      */
-    function price(uint32 challengeNumber) public view returns (uint256) {
+    function price(uint256 challengeNumber) public view returns (uint256) {
         Challenge memory _challenge = challenges[challengeNumber];
         if (_challenge.challenger == address(0x0)) {
             return 0;
